Move MetaStore and configuration storage to separate package.
