+++ server/src/main/java/io/atomix/copycat/server/storage/system/Configuration.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License
  */
-package io.atomix.copycat.server.storage;
+package io.atomix.copycat.server.storage.system;
 
 import io.atomix.catalyst.util.Assert;
 import io.atomix.copycat.server.state.Member;
